move bitrateTask to base
